requirements.txt
# Core
requests
beautifulsoup4
pandas

# OCR support
playwright
pillow
pytesseract

# LLM-based enrichment
openai

# Load .env files for local development (used by accident_info.py to read OPENAI_API_KEY etc.)
python-dotenv

# Google Drive integration
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

# pip install -r requirements.txt

# export OPENAI_API_KEY=".."
# source .venv/bin/activate

# python main.py "..."